requirements.txt
numpy
pandas
matplotlib
seaborn
torch
torchvision
scikit-learn
pillow
datasets
jupyterlab
ipykernel
pyyaml
wandb